added button for taking a new quiz, created new component for facts modal
