+++ client/src/pages/Results/ResultsFacts.js
@@ -1,0 +1,43 @@
+import React, { useState } from "react";
+import Button from "react-bootstrap/Button";
+import Modal from "react-bootstrap/Modal";
+
+function ResultsFacts(props) {
+  console.log(props);
+  const [show, setShow] = useState(false);
+  const handleClose = () => setShow(false);
+  const handleShow = () => setShow(true);
+
+  return (
+    <>
+      <Button
+        style={{ marginTop: "8%" }}
+        variant="primary"
+        onClick={handleShow}
+      >
+        Country Facts
+      </Button>
+
+      <Modal
+        show={show}
+        onHide={handleClose}
+        backdrop="static"
+        keyboard={false}
+      >
+        <Modal.Header closeButton>
+          <Modal.Title>{props.country}</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          Test
+        </Modal.Body>
+        <Modal.Footer>
+          <Button variant="secondary" onClick={handleClose}>
+            Close
+          </Button>
+        </Modal.Footer>
+      </Modal>
+    </>
+  );
+}
+
+export default ResultsFacts;
