+++ client/src/pages/Results/ResultsHeader.js
@@ -4,10 +4,7 @@ function ResultsHeader(props) {
   return (
     <div className="row text-center mr-3">
       <div className="col-sm-12 title-container p-3">
-        <h2>
-          {/* {questionState.title}  */}
-          {props.continent} Quiz Results
-        </h2>
+        <h2>{props.continent} Quiz Results</h2>
       </div>
     </div>
   );
